refactor: InputParser, InputValidator 모두 util 패키지로 이동
