+++ src/main/java/util/InputParser.java
@@ -1,4 +1,4 @@
-package domain;
+package util;
 
 import java.util.Arrays;
 import java.util.List;
