+++ src/main/java/util/InputValidator.java
@@ -1,9 +1,7 @@
-package domain;
+package util;
 
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 public class InputValidator {
     public static void validateInputNames(List<String> playerNames) {
