+++ src/test/java/util/InputParserTest.java
@@ -1,4 +1,4 @@
-package domain;
+package util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
