+++ src/test/java/util/InputValidatorTest.java
@@ -1,4 +1,4 @@
-package domain;
+package util;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
